feat(lib):增加`strntol`和`strntod` perf(lexer):减少一个环形缓冲区的参数 perf(lexer):优化fallthrough语句 perf(reader):简化`file_reader` fix(lexer):修复用del围住的多行字符串如果包含del会解析出错误的问题 fix(reader):修复产生结束符时长度计算错误的问题

+++ src/candy_wrap.c
@@ -60,7 +60,7 @@ void candy_wrap_append(candy_wrap_t *self, const void *data, int size) {
   void *dst = candy_wrap_get_data(self);
   if ((self->size + size) * candy_wrap_sizeof(self) > (int)sizeof(self->data)) {
     int limit = next_power2(self->size + size);
-    if (next_power2(self->size) < limit) {
+    if ((int)next_power2(self->size) < limit) {
       /* dst comes from either the stack space described by wrap::data or
          the heap space pointed to by wrap::data. if wrap::data is used to
          store the new heap space pointer directly, short data may be lost */
